Get the metadata of the meteorlogical element for a specific time resolution

The function getMetaElement() makes it possible to get the metadata of meteorological elements measured by met.no's stations

getMetaElement(timeserietypeID, stations = NULL)

Arguments

timeserietypeID

0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)

stations

id of the stations. Can be NULL, a scalar or a vector of station ids

Details

The output is a dataframe. This structure eases the conversion into json

Examples

tmp <- getMetaElement(timeserietypeID="2")
str(tmp)

#> 'data.frame':	261 obs. of  9 variables:
#>  $ description : Factor w/ 247 levels "\tJordens vanninnhold i 10 cm dypde",..: 158 155 194 168 220 226 220 227 220 228 ...
#>  $ element     : Factor w/ 261 levels "AA","BAT","BI",..: 104 107 112 116 187 188 189 190 191 192 ...
#>  $ elementGroup: Factor w/ 22 levels "Batterispenning",..: 10 10 14 14 18 18 18 18 18 18 ...
#>  $ elementNo   : Factor w/ 166 levels "-1","1","10",..: 19 24 20 52 121 123 124 125 126 127 ...
#>  $ fromDate    : Factor w/ 1 level "": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ language    : Factor w/ 1 level "no": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ name        : Factor w/ 188 levels "-","\tJordens vanninnhold",..: 121 118 146 145 170 171 170 171 170 171 ...
#>  $ toDate      : Factor w/ 1 level "": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ unit        : Factor w/ 19 levels "antall","åttedeler",..: 12 12 3 7 7 7 7 7 7 7 ...
#> 

tmp <- getMetaElement(timeserietypeID="0",stations=180)
str(tmp)

#> 'data.frame':	38 obs. of  9 variables:
#>  $ description : Factor w/ 38 levels "\"Timeverdi: Aritmetisk middel av minuttverdier Døgnverdi: Aritmetisk middel av 24 timeverdier (kl 00-00), evt formelbasert mid"| __truncated__,..: 27 15 4 36 8 22 14 1 19 17 ...
#>  $ element     : Factor w/ 38 levels "DD06","DD12",..: 26 14 27 24 12 15 32 25 28 30 ...
#>  $ elementGroup: Factor w/ 5 levels "Fuktighet","Lufttrykk",..: 4 1 4 1 5 2 4 4 4 4 ...
#>  $ elementNo   : Factor w/ 11 levels "-1","174","175",..: 1 1 1 1 1 2 8 5 6 7 ...
#>  $ fromDate    : Factor w/ 9 levels "1993-11-11T00:00:00.000Z",..: 7 6 1 6 5 5 8 1 5 5 ...
#>  $ language    : Factor w/ 1 level "no": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ name        : Factor w/ 37 levels "Blandingsforhold (mixing ratio)",..: 20 1 21 30 7 23 2 19 26 15 ...
#>  $ toDate      : Factor w/ 1 level "": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ unit        : Factor w/ 8 levels "gr/kg","graddager",..: 7 1 7 1 5 4 7 7 7 7 ...
#>