Get the metadata of the stations given several meteorlogical element

The function getMetaStationbyElement() makes it possible to get the metadata of the met.no's stations that measure all the chosen meteorological elements.

getMetaStationbyElement(timeserietypeID, elements)

Arguments

timeserietypeID

0:Daily, 1:Monthly, 2:Hourly, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)

elements

meteorlogical element as its code ("TAM", "RR", ...). Can be a vector of elements

Details

The output is a dataframe. This structure eases the conversion into json

Examples

tmp <- getMetaStationbyElement(timeserietypeID="2",elements="TA")
str(tmp)

#> 'data.frame':	947 obs. of  17 variables:
#>  $ amsl          : Factor w/ 421 levels "-1","0","1","10",..: 208 208 205 326 84 31 57 78 50 63 ...
#>  $ department    : Factor w/ 28 levels "","AKERSHUS",..: 9 9 9 9 19 19 2 2 2 2 ...
#>  $ fromDay       : Factor w/ 31 levels "1","10","11",..: 2 1 1 1 1 1 1 1 21 1 ...
#>  $ fromMonth     : Factor w/ 12 levels "1","10","11",..: 3 12 1 1 11 1 2 2 3 10 ...
#>  $ fromYear      : Factor w/ 118 levels "1833","1850",..: 96 76 60 60 68 60 91 94 109 75 ...
#>  $ latDec        : Factor w/ 891 levels "-1.0","-72.0117",..: 365 367 368 439 59 77 183 186 211 170 ...
#>  $ lonDec        : Factor w/ 899 levels "-1.0","-28.1",..: 241 240 238 224 189 191 184 185 194 181 ...
#>  $ municipalityNo: Factor w/ 300 levels "-1","1001","1002",..: 225 225 225 229 5 5 203 203 203 203 ...
#>  $ name          : Factor w/ 939 levels "16-11","AASEN",..: 864 863 862 123 616 82 296 294 31 295 ...
#>  $ stnr          : Factor w/ 947 levels "10000","10010",..: 66 85 92 572 7 30 122 123 136 175 ...
#>  $ toDay         : Factor w/ 5 levels "0","28","29",..: 1 5 5 1 4 5 2 5 1 5 ...
#>  $ toMonth       : Factor w/ 13 levels "0","1","10","11",..: 1 9 9 1 8 9 6 3 1 9 ...
#>  $ toYear        : Factor w/ 82 levels "0","1859","1872",..: 1 46 39 1 76 31 57 73 1 54 ...
#>  $ utm_e         : Factor w/ 907 levels "-10605","-114865",..: 525 524 523 515 451 456 452 454 468 449 ...
#>  $ utm_n         : Factor w/ 907 levels "0","6328707",..: 358 359 362 427 46 60 168 171 186 153 ...
#>  $ utm_zone      : Factor w/ 2 levels "0","33": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ wmoNo         : Factor w/ 393 levels "0","1001","1002",..: 301 301 301 300 391 1 390 390 378 390 ...
#> 

tmp <- getMetaStationbyElement(timeserietypeID="0",elements=c("TAM","RR"))
str(tmp)

#> 'data.frame':	926 obs. of  17 variables:
#>  $ amsl          : Factor w/ 419 levels "-1","0","1","10",..: 206 206 203 322 83 31 57 77 50 63 ...
#>  $ department    : Factor w/ 28 levels "","AKERSHUS",..: 9 9 9 9 19 19 2 2 2 2 ...
#>  $ fromDay       : Factor w/ 31 levels "1","10","11",..: 3 1 1 1 1 1 1 1 22 1 ...
#>  $ fromMonth     : Factor w/ 12 levels "1","10","11",..: 3 12 1 1 11 1 2 2 3 10 ...
#>  $ fromYear      : Factor w/ 114 levels "1837","1858",..: 91 71 55 55 63 55 86 89 105 70 ...
#>  $ latDec        : Factor w/ 891 levels "-1.0","-72.0117",..: 368 370 371 443 56 74 180 183 208 167 ...
#>  $ lonDec        : Factor w/ 899 levels "-1.0","-34.2103",..: 240 239 237 223 189 191 184 185 194 181 ...
#>  $ municipalityNo: Factor w/ 301 levels "-1","1001","1002",..: 225 225 225 229 5 5 203 203 203 203 ...
#>  $ name          : Factor w/ 918 levels "AASEN","ÅBJØRSBRÅTEN",..: 844 843 842 123 602 83 291 289 29 290 ...
#>  $ stnr          : Factor w/ 926 levels "10000","10010",..: 66 85 93 565 7 30 122 123 136 174 ...
#>  $ toDay         : Factor w/ 5 levels "0","28","29",..: 1 5 5 1 4 5 2 5 1 5 ...
#>  $ toMonth       : Factor w/ 13 levels "0","1","10","11",..: 1 9 9 1 8 9 6 3 1 9 ...
#>  $ toYear        : Factor w/ 82 levels "0","1859","1872",..: 1 46 39 1 76 31 57 73 1 54 ...
#>  $ utm_e         : Factor w/ 907 levels "-10605","-114865",..: 535 534 533 526 462 467 463 465 479 460 ...
#>  $ utm_n         : Factor w/ 907 levels "0","6328707",..: 360 361 364 433 44 58 167 170 185 151 ...
#>  $ utm_zone      : Factor w/ 2 levels "0","33": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ wmoNo         : Factor w/ 391 levels "0","1001","1002",..: 299 299 299 298 389 1 388 388 376 388 ...
#>