Get the values of one/several elements at one/several stations

The function getValues() makes it possible to get timeseries of meteorological elements at some given stations registered by met.no for a specific period of time and a specific time resolution. rem: timeserietypeID #2 is not always representing the hourly time resolution but also can be infra-hourly. We only kept hourly datasets.

getValues(timeserietypeID = NULL, fromPeriod = NULL, toPeriod = NULL,
  stations = NULL, elements = NULL, missingValues = -999)

Arguments

timeserietypeID

0:Daily, 1:Monthly, 2:Hourly, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)

fromPeriod

string of character yyyymmdd. Omitting this parameter will make it start 3 days earlier as today

toPeriod

string of character yyyymmdd. Omitting this parameter will make it stop today

stations

id of the stations. Can be a vector if multi-stations

elements

meteorlogical element as its code ("TAM", "RR", ...). Can be a vector if multi-element

missingValues

value of the expected mssing values. By default: -999

Value

The output is a dataframe structured as followed: the date timeserie, and the elements. Each element is a dataframe made of the timeseries of the different stations. This structure eases the conversion into a json format

Examples

########################################################
# Get daily temperature amd precipitation              #
# at station no 180 between 01/01/2013 and 05/01/2013  #
########################################################
tmp <- getValues(timeserietypeID="0",
                 fromPeriod="20130101",
                 toPeriod="20130105",
                 stations="180",
                 elements=c("TAM","RR"))
str(tmp)

#> 'data.frame':	5 obs. of  3 variables:
#>  $ date: Date, format: "2013-01-01" "2013-01-02" ...
#>  $ TAM :'data.frame':	5 obs. of  1 variable:
#>   ..$ 180: num  -1.1 -9.1 -8.7 -6 -4.6
#>  $ RR  :'data.frame':	5 obs. of  1 variable:
#>   ..$ 180: num  5.9 0 1.6 0.4 0.3
#> 

head(tmp)

#>         date  180 180
#> 1 2013-01-01 -1.1 5.9
#> 2 2013-01-02 -9.1 0.0
#> 3 2013-01-03 -8.7 1.6
#> 4 2013-01-04 -6.0 0.4
#> 5 2013-01-05 -4.6 0.3
#> 

########################################################
# Get hourly temperature at stations no 19710          #
# and 2650 between 27/12/2015 and 31/12/2015           #
########################################################
tmp <- getValues(timeserietypeID="2",
                 fromPeriod="20151227",
                 toPeriod="20151231",
                 stations=c("19710","2650"),
                 elements=c("TA"))
str(tmp)

#> 'data.frame':	120 obs. of  2 variables:
#>  $ date: POSIXct, format: "2015-12-27 00:00:00" "2015-12-27 01:00:00" ...
#>  $ TA  :'data.frame':	120 obs. of  2 variables:
#>   ..$ 19710: num  -6 -5.5 -5.3 -4.9 -3.8 -3.6 -3.7 -3.4 -4.4 -4.8 ...
#>   ..$ 2650 : num  -9.9 -10.1 -9.2 -8.6 -7.9 -6.7 -5.5 -3.8 -3.8 -3.9 ...
#> 

head(tmp)

#>                  date TA.19710 TA.2650
#> 1 2015-12-27 00:00:00     -6.0    -9.9
#> 2 2015-12-27 01:00:00     -5.5   -10.1
#> 3 2015-12-27 02:00:00     -5.3    -9.2
#> 4 2015-12-27 03:00:00     -4.9    -8.6
#> 5 2015-12-27 04:00:00     -3.8    -7.9
#> 6 2015-12-27 05:00:00     -3.6    -6.7
#> 

########################################################
# Get monthly temperature and precipitation at         #
# stations no 19710 and 180 between 16/03/2014 and     #
# 31/12/2015                                           #
########################################################
tmp <- getValues(timeserietypeID="1",
                 fromPeriod="20140316",
                 toPeriod="20151231",
                 stations=c("19710","180"),
                 elements=c("TAM","RR"))
str(tmp)

#> 'data.frame':	22 obs. of  3 variables:
#>  $ date: Date, format: "2014-03-01" "2014-04-01" ...
#>  $ TAM :'data.frame':	22 obs. of  2 variables:
#>   ..$ 19710: num  -999 6.8 11 14.8 19.6 14.9 12.3 8.3 3.3 -2.6 ...
#>   ..$ 180  : num  -999 3 8.3 12.2 17.8 12.8 8.8 5.6 0 -8.6 ...
#>  $ RR  :'data.frame':	22 obs. of  2 variables:
#>   ..$ 19710: num  -999 63.2 85.7 66.1 74 ...
#>   ..$ 180  : num  -999 53.9 83.9 46.4 106 ...
#> 

head(tmp)

#>         date TAM.19710 TAM.180 RR.19710 RR.180
#> 1 2014-03-01    -999.0  -999.0   -999.0 -999.0
#> 2 2014-04-01       6.8     3.0     63.2   53.9
#> 3 2014-05-01      11.0     8.3     85.7   83.9
#> 4 2014-06-01      14.8    12.2     66.1   46.4
#> 5 2014-07-01      19.6    17.8     74.0  106.0
#> 6 2014-08-01      14.9    12.8     99.4  152.9
#> 

########################################################
# Get daily normal temperature at                      #
# stations no 180, 2650                                #
########################################################
tmp <- getValues(timeserietypeID="4",
                 stations=c("180","2650"),
                 elements=c("TAM"))
str(tmp)

#> 'data.frame':	365 obs. of  2 variables:
#>  $ date: Date, format: "1961-01-01" "1961-01-02" ...
#>  $ TAM :'data.frame':	365 obs. of  2 variables:
#>   ..$ 180 : num  -10.7 -10.8 -10.8 -10.8 -10.9 -10.9 -10.9 -10.9 -10.9 -10.9 ...
#>   ..$ 2650: num  -7.7 -7.7 -7.7 -7.7 -7.8 -7.8 -7.8 -7.8 -7.8 -7.8 ...
#> 

head(tmp)

#>         date TAM.180 TAM.2650
#> 1 1961-01-01   -10.7     -7.7
#> 2 1961-01-02   -10.8     -7.7
#> 3 1961-01-03   -10.8     -7.7
#> 4 1961-01-04   -10.8     -7.7
#> 5 1961-01-05   -10.9     -7.8
#> 6 1961-01-06   -10.9     -7.8
#>